[Extensions] Add a BUILD.gn for the brailleDisplayPrivate API.
This CL adds a BUILD.gn file, and additionally creates a new
member function in ExtensionsBrowserClient on ChromeOS, to determine
if the current profile is active. This will be used in other APIs to
also break a dependency cycle on chrome/browser.

AX-Relnotes: n/a.
Bug: 40593486
Change-Id: I66df5fe1fdd92d32098ab9737823913119e64618
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5774370
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/main@{#1406276}