[ios] Fix Magic Stack card removal edge case
When removing a card as the NTP is going to disappear (but not being
deconstructed e.g. open in another tab), the card is removed from the
Magic Stack visually, but the space it took up is kept empty, and
upon returning to the NTP, swiping away and back brings that card back.
The proposed fix involves skipping the [UIView animateWithDuration:]
call and just executing the deleteItemFromDataSource block
synchronously if the removal is from a case that will result in a
navigation away from the NTP.

Fixed: 388916522
Change-Id: If717bf70773a7e52e8fdcaea2c712a3f033eab70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6168608
Reviewed-by: Scott Yoder <scottyoder@google.com>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406281}