[lensoverlay] Fix ghost loader disappearing on click
This CL changes the logic to close the ghost loader to only happen if
focus is no longer in the searchbox nor the ghost loader. Additionally,
I had to change the searchbox logic to keep querying for autocomplete
if the user clicked on the searchbox. Ideally the ghost loader is part
of the searchbox and this wouldn't be needed, but since this change
needs to be cherry picker for CSB M133 LE, doing the refactor is too
large a lift.

Finally, tabindex="-1" so that event.relatedTarget shows searchbox,
but we don't actually want ghost loader to be tab focus traversable,
hence the -1 value.

NO_IFTTT=I added the IFTTT for future use.

Bug: 383571713
Change-Id: I6f5cafe2bf216c35b4d61df3a04a1d88881ddf4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6168952
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1406322}