Return from HandleAttachmentToBrowserWindows when attachment is handled
This method loops over all windows, looking for a window to attach to or
to detach into a holder. However, once that happens we can exit but the
current code keeps looping, potentially executing the attachment
multiple times

This CL exits the method when attaching to a new window or the holder.
Additionally, `in_move_loop_` is now CHECKed to be false since the
method is always called after the move completes. This should now be the
case after https://crrev.com/c/6143033.

Bug: 377584872
Change-Id: If114f2d05c9c92a957abfbf417f997dff41853f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6171234
Reviewed-by: Andrea Gomez <andreaxg@google.com>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406330}