[Data Sharing] Fix solid color avatars for instant messages
Failing to set TINT_NONE for the ICON_TINT_COLOR uses a default tint
which masks the avatar [1]. This CL fixes that bug. The icon should
also be a circle. This CL adds corner radius the size of the avatar to
ensure that it appears as a circle.

[1]
https://source.chromium.org/chromium/chromium/src/+/main:components/messages/android/java/src/org/chromium/components/messages/MessageBannerProperties.java;drc=c0883e36f0f65273f002c2ca8b7e9474256e00e4;l=92

Fixed: 387509978
Change-Id: Iaa710876fbb97a50124a490a7c2a913e7c8242c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6175805
Reviewed-by: Sky Malice <skym@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406335}