media/gpu/chromeos: Duplicate FrameResource::tracking_token() methods
This works around a build dependency issue where the tracking_token()
method was undefined in libblink_module.so. When tracking_token is
defined in the base class, I get a linker issue due to a missing
dependency in the media/mojo/clients/BUILD.gn.

If I fix the build dependency issue, the linux-libfuzzer-asan-rel
builder finds a one definition rule (ODR) violation for
media::CodecPicture and perhaps other classes in media/gpu. The ODR
violation may be caused by an issue with Clank and inline methods.
Instead of working through these, it would be better for project
timelines to work around the issue.

Luckily, this is a temporary workaround. The commit can be reverted
after the CL that removes StableVideoDecoder
(https://crrev.com/c/5960256) lands. b/389986749 exists for that
cleanup.

Bug: b:376526494
Bug: b:389986749
Change-Id: Ice117086fdc305c6db703936c19691ed957cdf5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6173829
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406340}