Move Lens Overlay ActionItem invocation logic to the controller
This is a refactor with no intended changes in behaviour.

This CL moves the logic for invoking Lens Overlay's action item to its
entry-point controller, which handles logic for displaying the action
item already.

This is in preparation of a follow-up CL, where more logic will be
added, making `browser_actions.cc` a less suitable place for it.

Bug: 376283383
Change-Id: Idcf546cd9baa09978703eba4096fd7a9d16edbc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6172736
Reviewed-by: Darryl James <dljames@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406351}