OffscreenCanvasRenderingContext2D: Add getContextAttributes
After this was accidentally added to CanvasRenderingContext2D, it was
accidentally not also added to OffscreenCanvasRenderingContext2D.

In the IDL, this method should be in a common base class (in the
current spec draft this is CanvasSettings). This cannot be done yet
because only the OffscreenCanvasRenderingContext2D method is to be
gated on a feature. After this ships, and there is no more feature
flag, we can merge the two methods.

Tested locally (will be adding WPT tests only after the HTML spec
change is merged).

Bug: 388437261
Change-Id: I397863e61efafdc49ff0b0ace41c699072e6ccf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6169595
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406369}