Documenting when and where location_bar_view has null browser object
crbug.com/379534750 discussed extensively a ChromeOS-only crash and how
it can be mitigated. A follow-up bug b/380446664 was created to track
the future rework of simple_web_view_dialog to strip away
location_bar_view. In the meantime, a bug from a long time ago, from
ChromeOS Connectivity team, also documented the drawbacks of using
simple_web_view_dialog for the captive portal in the signin screen
during ChromeOS OOBE b/274707662.

Upon closer look, it appears that simple_web_view_dialog, which is
specifically for ChromeOS, is one of the only two classes that would
construct a location_bar_view with a null browser object in production
environment. The other one is presentation_receiver_window_view, which
applies to both ChromeOS and to other desktop platforms. It is also
worth noting that presentation_receiver_window_view is about to be
sunsetted in a year or so.

This CL documents extensively this learning.

Bug: 379534750
Test: tryjob
Change-Id: Ied54fd83910e6bbab7990679dd3f9f7a51426b72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6076090
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Di Wu <diwux@google.com>
Cr-Commit-Position: refs/heads/main@{#1406371}