[PWA] Refactor navigation capturing logic and ownership.
This CL combines all navigation capturing logic into one
NavigationCapturingProcess class, rather than having this be
split between various methods in web_app_launch_utils.cc and
the NavigationCapturingRedirectionThrottle class. This new
NavigationCapturingProcess class also serves the same purpose
as the old NavigationCapturingNavigationHandleUserData class,
in that it gets attached to the NavigationHandle and holds all
the data that is needed to make navigation capturing decisions
at redirect time.

Since NavigationCapturingProcess gets created before the
NavigationHandle to which it should be attached exists, a
templated helper NavigationHandleUserDataForwarder class is
also added. This class attaches the provided NavigationHandleUserData
to the first navigation to start in a given WebContents.

Finally, the launch metrics, launch queue and IPH handling is
separated from the navigation capturing logic/data in its own
WebAppLaunchNavigationHandleUserData class. For now this is
still tied to navigation capturing, but in the future we could
make this the way any launch parameters are attached to
navigations, and the way launch metrics are recorded for all
launches.

Bug: 351775835
Change-Id: I9a29b5fde067d92a0de088e178eb2174f213615c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6149709
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406384}