[Mac] Set task port policy crash key in all processes
The task port policy, particularly whether `amfi_get_out_of_my_way` is
set, is helpful for understanding certain types of crashes. It is now
set within `InitializeMac` within each process.

The return type of `GetMachTaskPortPolicy` is updated to separate the
success and error cases.

The system call is updated to use `__mac_syscall` directly.
`__sandbox_ms` is an alias of it. Since this isn't sandbox-related it's
clearer to use the generic syscall.

Change-Id: Id9e7870892f38522a0c3b30d3a681b8edff455e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6171814
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Mark Rowe <markrowe@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406431}