[spark][omnibox] `AutocompleteController::ShouldRunProvider()`
When the input is in the aggregator keyword mode, only run the
aggregator, keyword, and search providers. Similar to starter pack
suggestions.

When the input is not in the aggregator keyword mode, do not run the
aggregator provider. Similar to the @tabs starter provider.

Bug: 389078842
Change-Id: Ifc8734b25603928a8eafa0da6dd0b100a09ffa74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6164235
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406433}