[SIPool] Implement Create/DestroySharedImagePool() methods.
ClientSharedImageInterface adds the implementation of
SharedImageInterface::Create/DestroySharedImagePool() methods.
This in turn does a deferred IPC to the GPU process
via GpuChannel.

It is currently not implemented in the GPU service side.
It is also not used/plumbed to the SharedImagePool in the
client side as well.

Future CLs will do the implementation and plumbing.

Bug: 334764455
Change-Id: I74fd85b17db5f073d4cf2d1e1371863746fa6bdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6160073
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: vikas soni <vikassoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406478}