Force standalond display mode in IWAs
This commit adds overwriting "browser" and "minimal-ui" display modes in
IWAs to "standalone" and modifies the tests to check if that overwriting
works.

It also changes the HasMinimalUiButtons behaviour. Now it does not return true when the app is opened in the browser, and always returns true if the effective display mode is "minimal-ui".

Bug: 366524200
Change-Id: I4d47c9f91324536e8194c13b020cd3f945a3081e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6152057
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Patryk Chodur <pchodur@google.com>
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406568}