Show glic settings only if all flags enabled
Before this the settings page would show if the kGlic feature is enabled
but the prefs are exposed to the settings page (in prefs_util.cc) based
on GlicEnabling::IsEnabledByFlags which is true only if both kGlic and
kTabstripComboButton are set. This was leading to pref errors on the
settings page if Chrome was started with only kGlic enabled.

This CL bases Glic settings visibility on GlicEnabling as well.

Bug: 379166610,386362596
Change-Id: Ic6aaff8e37aaf0f9013f7c3265bd03b64cae0844
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6171235
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406639}