[Desktop][iOS] Use ProfileSignout::kUserDisabledAllowChromeSignIn
In crrev.com/c/5683002, a new signout reason,
kUserDisabledAllowChromeSignIn, was introduced for allow chrome sign-in
in google services.

kUserDisabledAllowChromeSignIn is already used on Android.

This CL updates desktop, as suggested in https://chromium-review.googlesource.com/c/chromium/src/+/5683002/comments/7e43327a_34488d71, and iOS to use this reason.

Note: The histogram Signin.SignOut.Completed is expected to change with
this CL:
kUserDisabledAllowChromeSignIn will be recorded for desktop and iOS
instead of kUserClickedSignoutSettings on iOS, and
kSigninManagerUpdateUPA on desktop.

Bug: 350699437, 375605174
Change-Id: I7fac13529ed85c767884b5a90cc8ec5898d75028
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6171278
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406656}