Fix origin for VideoResourceUpdater::CopyHardwarePlane
Historically it was using origin from the source shared image, while it
sends a copy to the display compositor. Destination for copy is created
using top left and CopySharedImage call will perform flip during it.

This CL fixes to use destination origin.

This CL is no-op, because the only case when we need frames to be
copied is hardware decoded frames and they are always top-left [1].
Frames that are created from webgl are already accessible cross-thread
and don't need the copy.

[1] https://source.chromium.org/chromium/chromium/src/+/main:media/gpu/android/direct_shared_image_video_provider.cc;drc=8d4fd94dbbf78e0aae68a8e99a2316d1bb63bf44;l=193

Bug: 378688985
Change-Id: Ib2e09988bcf4ad124f58b9612a27043570ab3ef8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6173173
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406658}