Add flag allowing to skip network setup for testing
This is useful in factories that have no internet access and need to use
guest mode. With recent changes, this is no longer possible from the
initial network screen due to the requirement to connect to the network
and obtain the initial enterprise state. It is possible to disable this
logic using the following command line switch:

  --enterprise-enable-unified-state-determination=never

Unfortunately, this switch is only interpreted after we reach the
auto-enrollment-check screen, but users without an internet connection
get stuck at the network-selection screen and never reach it. The switch
added in this CL allows to bypass the network setup and proceed through
the rest of OOBE.

Test: add command line switches `--enterprise-enable-unified-state-determination=never` and `--oobe-skip-network-setup-for-testing=true` to /etc/chrome_dev.conf, disconnect internet, proceed through OOBE and enter guest mode
Bug: b:384377276
Change-Id: I6b54af73f1d59ac4803348b744443ecf79fa8cf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6152059
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Sergiy Belozorov <sergiyb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406673}