android: Restrict PdfPageTest not to run on Automotive
This test fails with "PDF document is not loaded successfully". I am 99%
sure loading PDFs in Chrome is not supported on Automotive yet.

The test was excluded for this bot via restriction @MinAndroidSdkLevel
in this cherry-pick: https://crrev.com/c/6175114.

This commit does not fix any failures, just indicates that PDF support
is not there yet on all tested configurations for Automotive.

Bug: 389932061, 389959412
Change-Id: I3c0e66937b9dfa36d1509269c6d2897c3f07c18a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6169000
Reviewed-by: Shu Yang <shuyng@google.com>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406678}