[Protected Audiences] Make trusted signals KVv2 use "renderURLs"
As well as "componentRenderURLs". It was sending and expecting to
receive "renderUrls" instead.

The new capitalization matches the value that's exposed to scripts,
and matches the common style used by web Javascript APIs.

The only server implementation supports both capitalizations, so this
is safe to change. There can be no other server implementations, due
to the keys for the extra layer of encryption on top of HTTPS, so no
need to worry about other implementations existing.

Fixed: 389879225
Change-Id: I4431499180ff095036f72e36e6915a13a4aa80a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6169484
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406748}