[OmniboxAPI] Support OnInputEntered
The OnInputEntered event is currently only sent to matches with keyword
that correspond's to a template URL keyword. This CL sets
match.keyword for unscoped suggestions to the keyword to be able to
retrieve the template url (and hence the extension  ID), and forward
the match selection to the extension that previously made the suggestion.

Fixed: 381280376
Change-Id: If6c64bcce2ce2a09cf3277431dcd1085247258d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6149668
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Cr-Commit-Position: refs/heads/main@{#1406790}