Make downloads pinnable to the toolbar (part 5)
Add the downloads progress ring as an installable view so that it can be
added to the generic ToolbarButton created for the downloads action
item. The progress ring has a dormant, scanning, and downloading state.

Bug: 323962334
Change-Id: Ifa126a608963c7445412ad04b5340e11fa7c5e71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6173677
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406798}