[Messaging] Notify locally generated tab events to the service
This CL adds plumbing to propagate locally generated events
for tabs and tab groups through the notifier to its observers.
Functionally this CL is a no-op. In the next CL, we will add
functionality that should be taken on the service (e.g. clearing
dirty messages) in response to these locally generated events.

Bug: 385003046
Change-Id: I89637d61eb212cac6a84027169d141a2285df75c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6174600
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406801}