Avoid PseudoStateChanged call during HTMLInputElement construction.
DidNotifySubtreeInsertionsToDocument() is called during the process of
inserting an element into the document.  Prior to this change
HTMLInputElement's implementation calls ListAttributeTargetChanged().
This change makes it instead only do one of the two things that
ListAttributeTargetChanged() does, since the other
(PseudoStateChanged()) is unnecessary since there has been no *change*
to the pseudo state.  This avoids the cost of computing the style
changes involved.

This was found in a profile of the Speedometer3 subtest
TodoMVC-jQuery/CompletingAllItems, on which it is a 2% win.  It also
appears to be about a 0.5% win on TodoMVC-Javascript-ES5 and
TodoMVC-Javascript-ES6-Webpack-Complex-DOM (where it also shows up in
the profiles).

Change-Id: Ie76c13b6fe6eccb421a78c388582c7f5a3b4662a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6174303
Reviewed-by: Di Zhang <dizhangg@chromium.org>
Commit-Queue: Di Zhang <dizhangg@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406805}