[TabGroups] Remove observer before TabGroupSyncService is destroyed
When the TabGroupSyncService is destroyed it will clear any user data
that is attached to it.

An MSAN failure happens in the case where the TabGroupSyncService is
destructed and the TabGroupTypeObserver continues to listen to the
service. In this case, the memory is freed before we have a chance to
properly Remove TabGroupTypeObserver as an observer of
TabGroupSyncService.

To prevent this we override TabGroupSyncService::OnWillBeDestroyed.
This is called right before ClearAllUserData() is called. Allowing us
to stop listening to TabGroupSyncService before it goes out of scope.

Change-Id: I36e7dc25a2e107199dd57511dd949c8c2e35a430
Bug: 389737049
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6172815
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406810}