firmware_update: check for restart on update complete
While crrev.com/c/6027313 successfully prevents the restart button from
showing on the failure dialog, manual testing revealed that the restart
buttons no longer show on the success dialog. This is because of the
added check, `this.installationProgress.state ===
UpdateState.kSuccess`. `updateRequiresRestart` is computed on template
load and not recomputed when the update completes, so, it will always be
false. This was not an issue before `crrev.com/c/6027313` as
`update.needsReboot` was available when `updateRequiresRestart` was
computed. This was not caught in tests as `InstallationProgress` was
immediately set to `UpdateState.kSuccess.` Upon changing the tests to
first `setInstallationProgress(1, UpdateState.kUpdating)`, we see that
the previous code actually failed to exhibit expected behavior.

Now, `updateRequiresRestart` will be recomputed when `updateIsDone`
changes. I have also changed the tests to first set the state to
`kUpdating` before testing the behavior they want to see.

Test: out/Default/browser_tests --gtest_filter=FirmwareUpdateAppBrowserTest.FirmwareUpdateDialog

Test: manually ensure the restart buttons appears on successful
updates.

Test: manually ensure the restart buttons do not show up on
the failure dialog by forcing a failed update, e.g. by making
`isUpdateDone` set `installationProgress.state = UpdateState.kFailed`

Bug: b:379342936

Change-Id: Id5b2a9254894112928b6f444c39313661d5efacf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6167946
Reviewed-by: Michael Checo <michaelcheco@google.com>
Commit-Queue: Andre Braga <andrebraga@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Cr-Commit-Position: refs/heads/main@{#1406818}