Introduce helper function detecting target Toolbar placement.
The logic assumes the request is supplied for a tab that does not have
any of its editable fields focused (including web forms) and is not
triggered as a result of a preference change.

The change caches the value of SharedPreference to avoid frequently
reading the SharedPreferences content.

Change-Id: Id545f9fdf1815cfa19d231bd5308ad8e5027e53f
Fixed: b:388526642
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6175382
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1406829}